requirements.txt
# 视觉大模型教程 - 生产环境依赖
# Python >= 3.8

# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0

# 视觉大模型
clip @ git+https://github.com/openai/CLIP.git
pillow==10.1.0
opencv-python==4.8.1.78

# 数据处理
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2

# 图像处理和增强
albumentations==1.3.1
imgaug==0.4.0

# 模型微调
peft==0.6.2
bitsandbytes==0.41.3
accelerate==0.25.0

# 模型部署
onnx==1.15.0
onnxruntime==1.16.3
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# 配置和日志
pyyaml==6.0.1
python-dotenv==1.0.0
loguru==0.7.2

# 可视化
matplotlib==3.8.2
seaborn==0.13.0
tqdm==4.66.1

# Jupyter支持
ipywidgets==8.1.1

# HTTP请求
requests==2.31.0
aiohttp==3.9.1

# 数据集下载
huggingface-hub==0.19.4
datasets==2.15.0

# 工具
fire==0.5.0

